iraffunctions: Take out some test code so binary format is used for save files. Modify clProcedure to give better name to CL scripts executed via 'cl < filename.cl'. Modify iraftask so it uses these cleaner names too.
minmatch: Defer creation of minmatch dictionary until it is needed. Add a few other minor speedups.  This is almost twice as fast (when reading save files, at least.)

irafpar: defer creating the list of pset parameters until it is needed.  This fixes a bug in the save/restore procedure.

irafhelp: Add methods as separate group of objects.  Modify format for module listings (now a table).  Add sizes of lists & dictionaries to help info.


git-svn-id: https://aeon.stsci.edu/ssb/svn/pyraf/trunk@235 05fa7843-0683-0410-9e38-dfee850dc5be