Make IRAFARCH variable optional
This variable is unused f.e. in the Debian/Ubuntu IRAF package, and
setting it will confuse the (newer) installation, beause it is used as
suffix for the bindirs.

If IRAFARCH is unset, we can just use the plain bin/, unix/bin/ and
noao/bin/ directories that are usually created during an installation.