Fix bug in clExecute argument list
The argument list used and empty dict resp. an empty list as default
for local_vars_dict resp. local_vars_list. This is a problem, since
they may be changed within the function (they are then used in the
call of cl2py()), and the change remains global. Since cl2py accepts
None here (and uses an empty list implicitly), we can safely switch to
None as defaults here.