Provde a script to fill the system clcache
This script is a useful admin tool to pre-fill the system cache with
all precompiled scripts from the IRAF system. Since it is of special
use, it is somehow hidden and can be invoked with

    $ python3 -m pyraf.fill_clcache

The tools is quite simple as it just searches the IRAF path for `.cl`
scripts and compiles them. Errors are just ignored.