First test for result code in test_invocation
This reorders the asserts a bit so that first the result code from the
invocation is checked (and on error, stderr is returned), and then the
output is checked. This has the advantage that we may see why the
invocation failed instead of just seeing that stdout was empty.

It is mysterious however why sometimes the invocation fails (as seen
on Ubuntu, armhf).