Simplify FileCache
`FileCache` comes with a "base" implementation that caches the file
content. This is however always overwritten in the child classes, and
`FileCache` is never used directly; we can just remove this "base"
implementation.

`MD5Cache` uses in its implementation the reading of the base class
the `_getFileHandle()` method of the base class; this is the only
place where this method is used. So, we could move it to `MD5Cache`,
but we could also just pragmatically replace it with a simple binary
`open()` function (and properly handle it in a `with` clause).

FileCache._getAttributes method is never used with an argument, and
also not with a file handle. The code which handles this is removed as
well.